+++ src/main/java/dev/willyelton/crystal_tools/common/config/CrystalToolsConfig.java
@@ -73,7 +73,6 @@ public class CrystalToolsConfig {
 
     // Backpack
     public static ModConfigSpec.EnumValue<CrystalBackpackInventory.SortType> BACKPACK_SORT_TYPE;
-    public static ModConfigSpec.IntValue BACKPACK_BASE_EXPERIENCE_CAP;
     public static ModConfigSpec.IntValue MAX_COMPRESSION_SLOT_ROWS;
 
     // Trident
@@ -191,8 +190,6 @@ public class CrystalToolsConfig {
         builder.comment("Crystal backpack settings").push("backpack");
         BACKPACK_SORT_TYPE = builder.comment("Method used for sorting the Crystal Backpack")
                 .defineEnum("backpack_sort_type", CrystalBackpackInventory.SortType.QUANTITY);
-        BACKPACK_BASE_EXPERIENCE_CAP = builder.comment("Starting EXP Cap for the Backpack")
-                .defineInRange("backpack_base_experience_cap", 200, 1, 10000);
         MAX_COMPRESSION_SLOT_ROWS = builder.comment("Maximum number of rows of compression slots. These slots will not currently scroll so don't set it to larger than your gui scale can render")
                 .defineInRange("max_compression_slot_rows", 6, 1, 20);
         builder.pop();