+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/tool/CrystalShield.java
@@ -71,11 +71,6 @@ public class CrystalShield extends ShieldItem implements LevelableItem, EntityTa
     }
 
     @Override
-    public String getItemType() {
-        return "shield";
-    }
-
-    @Override
     public boolean isDisabled() {
         return false;
     }