+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/tool/DiggerLevelableTool.java
@@ -9,12 +9,10 @@ import dev.willyelton.crystal_tools.utils.ToolUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 import net.neoforged.neoforge.common.Tags;
@@ -27,8 +25,8 @@ import java.util.function.Predicate;
  * Handles the 3x3 mining and vein mining
  */
 public abstract class DiggerLevelableTool extends LevelableTool implements VeinMinerLevelableTool {
-    public DiggerLevelableTool(Item.Properties properties, String toolType) {
-        super(properties, toolType);
+    public DiggerLevelableTool(Item.Properties properties) {
+        super(properties);
     }
 
     @Override